refactor: update clock imports
