+++ src/components/views/audio_messages/DurationClock.tsx
@@ -8,7 +8,7 @@ Please see LICENSE files in the repository root for full details.
 
 import React from "react";
 
-import Clock from "./Clock";
+import { Clock } from "../../../shared-components/audio/Clock";
 import { type Playback } from "../../../audio/Playback";
 
 interface IProps {
