+++ src/components/views/messages/LegacyCallEvent.tsx
@@ -18,7 +18,7 @@ import { LegacyCallEventGrouperEvent } from "../../structures/LegacyCallEventGro
 import AccessibleButton from "../elements/AccessibleButton";
 import InfoTooltip, { InfoTooltipKind } from "../elements/InfoTooltip";
 import { formatPreciseDuration } from "../../../DateUtils";
-import Clock from "../audio_messages/Clock";
+import { Clock } from "../../../shared-components/audio/Clock";
 
 const MAX_NON_NARROW_WIDTH = (450 / 70) * 100;
